+++ src/main/java/lotto/option/LottoOption.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.option;
 
 import lotto.number.WinningNumbers;
